Fixed the Code
I fixed the code and did a quick test on the Q to check. The problem was in the thread creation conditional. The way it was written, it would break out of the for loop if the first thread was created properly. As a result, if the first thread was created correctly the second one would never be made.